[Main group 6][Guilherme][Back-end][products routes]

+++ back-end/src/database/migrations/20220714191706-create-product.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.STRING(100)
       },
       price: {
-        type: Sequelize.DECIMAL(4, 2)
+        type: Sequelize.FLOAT
       },
       url_image: {
         type: Sequelize.STRING(200)
